[Comgr][hotswap] Split the register state out of the raise context
Most of what the raise context carried was about registers: the alloca-backed
register file and its entry layout, the operand decode that resolves a register
operand, the reads and writes that go through it, EXEC predication of the
per-lane writes, and the per-SGPR shadow storage that carries wave masks and
source-wave values across block boundaries. All of it sat next to the block
map, the source text bytes and the private segment allocation, which have
nothing to do with registers.

Give that half its own home. RegisterState holds the register file, the
user-SGPR layout and every fact keyed by a register, and takes the builder, the
projection and the MC layer directly, so the dependency runs one way: the
register state reaches the projection, the context reaches the register state,
and nothing reaches back. Handlers get there through RaiseContext::registers().

What stays is what is not register-keyed: the block map, the source code object
the PC-relative literals are read out of, the source private segment
allocation, and the lane and wave semantics the projection defines.

The tests follow the code they cover: everything exercising registers moves to
a file named after the register state, and what is left pins the context
itself, its block lookup included.
