+++ amd/comgr/src/hotswap/decoder/parsed-reg.h
@@ -23,7 +23,7 @@ struct ParsedReg {
   //   SRC_SCC  : i1 == SCC
   // These cannot be written. We model them as their own kinds so
   // readOp32 / readOp64 can materialise the boolean result on demand
-  // (see raise-context.cpp). Tensile gfx1250 emits them as F16 source
+  // (see register-state.cpp). Tensile gfx1250 emits them as F16 source
   // operands (e.g. `v_sub_f16 v64, src_vccz, v48`), so the dispatch
   // path must recognise them or the kernel crashes inside parseReg.
   // VCC_HI_SCRATCH / EXEC_HI_SCRATCH: on a WAVE32 source, hardware VCC and
